+++ integration-tests/static-buffer/lib.rs
@@ -2,6 +2,11 @@
 
 #[ink::contract]
 pub mod static_buffer {
+    use ink::prelude::{
+        string::String,
+        vec::Vec,
+    };
+
     #[allow(unused_imports)]
     use ink::env::BUFFER_SIZE;
     #[ink(storage)]
@@ -27,6 +32,19 @@ pub mod static_buffer {
         #[ink(message)]
         pub fn get_caller(&self) -> AccountId {
             self.env().caller()
+        }
+
+        #[ink(message)]
+        pub fn buffer(&self) -> Result<(u64, u64), String> {
+            let buf1 = Vec::<u8>::with_capacity(3);
+            let buf2 = Vec::<u64>::with_capacity(1);
+            let ptr1 = buf1.as_ptr() as u64;
+            let ptr2 = buf2.as_ptr() as u64;
+            let align = core::mem::align_of::<Vec<u64>>() as u64;
+            let padding = ptr2
+                .checked_sub(ptr1)
+                .ok_or(String::from("Error during padding calculation"))?;
+            Ok((padding, align))
         }
     }
 
@@ -73,6 +91,33 @@ pub mod static_buffer {
                 super::BUFFER_SIZE.to_string()
             );
 
+            Ok(())
+        }
+
+        #[ink_e2e::test]
+        async fn buffer<Client: E2EBackend>(mut client: Client) -> E2EResult<()> {
+            // given
+            let mut constructor = StaticBufferRef::new_default();
+
+            // when
+            let contract = client
+                .instantiate("static_buffer", &ink_e2e::bob(), &mut constructor)
+                .submit()
+                .await
+                .expect("instantiate failed");
+            let call_builder = contract.call_builder::<StaticBuffer>();
+
+            // then
+            let buffer_call = call_builder.buffer();
+            let buffer_call_res =
+                client.call(&ink_e2e::bob(), &buffer_call).submit().await?;
+            let value = buffer_call_res.return_value();
+            assert!(value.is_ok());
+            let value = value.unwrap();
+            let padding = value.0;
+            let align = value.1;
+            assert_eq!(padding, 8);
+            assert_eq!(align, 4);
             Ok(())
         }
     }